fixing shopping cart and shopping bag

+++ src/tests/AccaptanceTests/StoreTests/DiscountPolicyTest.java
@@ -24,7 +24,7 @@ public class DiscountPolicyTest extends ServiceTests {
     public void testAddMinQuantityDiscountPolicySuccess() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertTrue(addMinQuantityDiscountPolicy(sessionID1, storeID, 0, productID1, 2, false) == dId);
     }
 
@@ -32,7 +32,7 @@ public class DiscountPolicyTest extends ServiceTests {
     public void testAddMinBagTotalDiscountPolicySuccess() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertTrue(addMinBagTotalDiscountPolicy(sessionID1, storeID, 0, 6) == dId);
     }
 
@@ -40,7 +40,7 @@ public class DiscountPolicyTest extends ServiceTests {
     public void testAddMinBagTotalDiscountPolicyStoreNotExistFail() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long dId = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertFalse(addMinBagTotalDiscountPolicy(sessionID1, storeID+2, 0, 6) == dId);
     }
 
@@ -48,21 +48,21 @@ public class DiscountPolicyTest extends ServiceTests {
     public void testAddMinBagTotalDiscountPolicyDiscountNotExistFail() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertFalse(addMinBagTotalDiscountPolicy(sessionID1, storeID, -1, 6) == did);
     }
     @Test
     public void testAddMinBagTotalDiscountPolicyNegativeMinQuantityFail() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertFalse(addMinBagTotalDiscountPolicy(sessionID1, storeID, 0, -1) == did);
     }
     @Test
     public void testAddMinQuantityPurchasePolicyFailure() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         assertFalse(addMinQuantityDiscountPolicy(sessionID1, storeID + 1, 0, productID1, 2, false)== did);
         assertFalse(addMinQuantityDiscountPolicy(sessionID1, storeID, 1, productID1, 2, false) == did);
         assertFalse(addMinQuantityDiscountPolicy(sessionID1, storeID, 0, productID1 + 1, 2, false) == did);
@@ -73,7 +73,7 @@ public class DiscountPolicyTest extends ServiceTests {
     public void testAndPoliciesSuccess() {
         int storeID = openStore(sessionID1, "newStore");
         int productID1 = addProduct(sessionID1, storeID, "test", 3.9, "milk", 9, "10");
-        int did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
+        long did = addProductDiscount(sessionID1, storeID, productID1, 0.1, 0);
         Integer p1 = addMinBagTotalDiscountPolicy(sessionID1, storeID, 0, 3);
         Integer p2 = addMinBagTotalDiscountPolicy(sessionID1, storeID, 0, 6);
         assertFalse(p1!=null && p2 !=null);
